割り算の間違いを修正

+++ q10251743340/ソースファイル２.cpp
@@ -2,7 +2,7 @@
 double AVE_MIN_MAX;
 double ave3(int a, int b, int c) {
 	double ave;
-	ave = (double)(a + b + c) / 3;
-	AVE_MIN_MAX = (double)(MIN3 + MAX3) / 2;
+	ave = (a + b + c) / 3.;
+	AVE_MIN_MAX = (MIN3 + MAX3) / 2.;
 	return ave;
 }
